Address more review comments
