+++ new-integration-tests/src/test/resources/wdt-models/model-singlecluster-two-sampleapp-wls.yaml
